requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Django==4.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
inflection==0.5.1
ipython==8.20.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==8.1.0
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.13
